requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
aiogram==2.25.1
SQLAlchemy==2
alembic==1.11.1
apscheduler==3.10.1
python-dotenv==1.0.0
httpx==0.24.1
pydantic==1.10.8
python-multipart==0.0.6
